feat(generate): read css properties defined on host
