+++ generate/queries/css/cssCustomProperties.scm
@@ -1,4 +1,8 @@
-(
+( ; color: /** blue */ var(--blue);
+  ; color: light-dark(/** blue */
+  ;                   var(--blue),
+  ;                   /** dark blue */
+  ;                   var(--dark-blue))
   (comment)? @comment (#match? @comment "^/\\*\\*")
   (call_expression
     (function_name) @fn (#eq? @fn "var")
@@ -8,5 +12,16 @@
       (plain_value) @property (#match? @property "^--[^_]")
       ("," [_(_)]* @default)?
       .
-      ")"))) @cssPropertyCallSite
+      ")"))) @cssProperty
 
+( ; /** defined on host */
+  ; --var: blue
+  (rule_set
+    (selectors
+      (pseudo_class_selector
+        (class_name) @class.name (#eq? @class.name "host")))
+    (block
+      (comment)? @comment (#match? @comment "^/\\*\\*")
+      (declaration
+        (property_name) @property (#match? @property "^--[^_]")
+        (_)* @default)))) @cssProperty
