--- generate/style_cache.go
@@ -1,48 +1,0 @@
-/*
-Copyright © 2025 Benny Powers
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program. If not, see <http://www.gnu.org/licenses/>.
-*/
-package generate
-
-import (
-	"sync"
-)
-
-// CssParseCache caches parsed CSS files by absolute path.
-type CssParseCache struct {
-	mu    sync.RWMutex
-	cache map[string]CssPropsMap
-}
-
-func NewCssParseCache() *CssParseCache {
-	return &CssParseCache{
-		cache: make(map[string]CssPropsMap),
-	}
-}
-
-// Get checks the cache for a parsed result.
-func (c *CssParseCache) Get(path string) (CssPropsMap, bool) {
-	c.mu.RLock()
-	defer c.mu.RUnlock()
-	val, ok := c.cache[path]
-	return val, ok
-}
-
-// Set stores a parsed result.
-func (c *CssParseCache) Set(path string, props CssPropsMap) {
-	c.mu.Lock()
-	defer c.mu.Unlock()
-	c.cache[path] = props
-}
