+++ generate/test/fixtures/project-css/.config/cem.yaml
@@ -1,0 +1,3 @@
+sourceControlRootUrl: https://github.com/bennypowers/cem/tree/main/generate/test/fixtures/project-css/
+files:
+  - src/*.ts
