+++ generate/test/fixtures/project-css/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "project-css",
+  "customElements": "custom-elements.json"
+}
