+++ generate/test/fixtures/project-css/src/class-css-host-props.ts
@@ -1,0 +1,5 @@
+import styles from './class-css-host-props.css'
+@customElement('class-css-host-props')
+class ClassCssHostProps extends LitElement {
+  static styles = styles;
+}
